+++ Changelog.md
@@ -14,6 +14,8 @@ and this project adheres to
 -   Confirm dialog before app closes while profiling
 -   Confirm dialog before app closes while data is processing
 -   Tooltips for more UI Components
+-   Handle error when processing "Battery voltage does not cross the defined low
+    cut off voltage. Please define higher cut off level and run again."
 
 ### Fixed